execute external git calls using subprocess
this has the advantage of not spawning a shell, eliminating potential
problems with whitespace in path/file names.

as opposed to the original implementation, this *will* raise an
exception if the git call fails. for lack of knowledge on how calypso
prefers to handle exceptions, it goes unhandled for now.