fix double items on put requests
the call to scan_file in Collection.create_file without adding the
file to Collection.files made the following scan_dir create a
duplicate item, which persisted until a change of the item or a
restart of calypso.

fixed by removing the call. it will be inserted into items by scan_dir
on its own anyway.