Use short timeouts when reading first HTTP line to deal with CalDAV-Sync
For some reason, CalDAV-Sync doesn't reliably close the socket, so we
want to use a short timeout around that, but we want a longer timeout
for the rest of the socket uses. Poke the socket timeouts around the
spot where the first line is read and reset them to a much longer time
afterwards.

Signed-off-by: Keith Packard <keithp@keithp.com>