Access to the root directory or any other non-leaf directory raises a IOError
IOError: [Errno 21] Is a directory: '.../.config/calypso/calendars/...'

which is caught and logged as an exception, but still shows even though
just a file-oriented function is called in scan_dir with a subdirectory.
Instead, now, a debug-level log message is emitted on directories that
they are ignored for the purpose of scan_dir.